Adding downloads bar, miner headerbar improvements.
The idea behind the downloads bar design is to reduce perceived complexity. It looks simpler than a popover or dialog due to it's relatively static nature.

Also added more padding to the addressbar to give more space to grab the window by and changed the event during which the back and forward buttons' sensitivity is updated so that users can navigate back while a page is loading.