Fix broken viewsource.
This broke in the last commit when I fixed pinning to be restored
AFTER the tab is added, because at that time it 'restored' the URL
back to odysseus:home.

Besides I think sending the URL back to the BrowserWindow for display
is cleaner, though I don't like the new global method.