Correct problems with launching Odysseus + simplification.
The issue was that when Odysseus opens it couldn't open any URLs,
as open() was called before initialize() thereby creating null pointer exceptions.

During that work I also removed my attempt at not creating a new window
when one was created anyways, as the logic for that was messy and not
worth it.