Autocomplete URIs from browser history.
Includes infrastructure for implementing autocompleters in SQL.
Adds deduplicating of autocompletions
(to avoid repeating completions within history autocompletion and
between autocompleters).

It is worth noting the autocompletion infrastructure is now essentially
an implemention of a DISTINCT UNION, but allowing for online databases
to be queried via whatever HTTP API they expose.