Get {% query %} to work with {% macro %} (as well as {% with %} & {% trans %}.
These all use the same runtime node with their real implementation
being done at compile time (except for {% with %} which lets you
instantiate the AST node directly).

The main challenge here was around inlining variables within others.
For that I defined a new method on variables which'll try and
detect whether it actually needs to build a new one (to save memory),
before it copies over the data from the variable context and extend it
to have the new filters (with recursive inlining) and path traversal
(implemented as a behind-the-scenes filter).

Simplified Odysseus.Templating.Variable.nilvar initialization.