Communicate which tab was used for history navigation.
This was already communicated to screenreaders quite well,
by allowing them to navigate through records not just chronologically
but by where the surfer came to a page from.

But now I've added colours to the times to communicate which tab was used.
This required a custom filter and non-obvious code to keep the colours
nicely spread out.