Draft code for adding recommendations to the homepage.
I'll write this for realsies (probably via PhantomJS & maybe a shell script
for converting to SQLite files as that'll be easier to integrate to odysseus:home)
when I've otherwise got a nice topsites display working on odysseus:home.

And then I'll integrate it there. At that time there'd be a question of how
to distribute this database.