Fix and enable the fetch tag.
Previously I tried to hold the rest of the template in the fetch AST,
but that doesn't play nicely with other block tags.
And besides I didn't need to do that to hold up rendering of
the rest of the page.

Fixing that had several flow-on bugs to fix.

And I disabled handling of XML responses for now.