Move away from writing my own URI escaper, I don't trust it well enough.
The tricky bit here is that all other autoescapers are data driven, not logic driven.
And as such I needed to introduce a way for variables to call some code
to perform it's autoescaping rather than it's more scalable data-driven logic.

So I let those escapers point to a filter to use via the otherwise invalid
\0 substitution/escape entry.
Then the Variable constructor corrects this to maintain it's existing constraints.