Tell webservers what formats {% fetch %} can parse, allow permacaching.
(optional) permacaching will be useful as an autoupdate for browser addons,
which no matter how much RMS dislikes it (calling it a "universal backdoor")
is vital as if not updated this software may fully break.

This is hidden behind a flag because if it's not asked for it will
often just be a waste of diskspace.