Enable, integrate, and fix AppStream-based app suggestions.
Downloads now make sure you have apps to open them with.
Invalid URIs now report errors better,
and failing to them will also suggest some apps.
And when I have an RSS subscription UI, it will have a page to link to
in order to ensure you have an app to subscribe with.