Improve performance of odysseus:home.
This is necessary as the first thing you'll probably want to do on this page is
to enter a URL or web search. But the freeze could cause people to question if
they can do that.

I arrived at the two queries I'm using by trial and error in the SQLite shell,
but my suspicions on why they work well is that:
1. The historical_locations index makes counting up visits fast.
2. Explicitly not inserting entries already present in the new cache table
    tells SQLite not to waste time inserting those into the url_by_count index.
3. The url_by_count index gives everything needed to make the main query fast,
    and caching it on-disk minimizes the reprocessing to make this work.

This struggle nicely illustrates a failure point of most RDBMS optimizers/
query-planners. They don't know how frequently used a query will be and
hence they can't optimize them by caching partial query results on-disk.

Database admins have to explicitly add indexes.