Don't warn about missing translations.
There's just too many of these warnings & they don't mean anything.

I tried to fix this by detecting whether the text is an english translation,
but I still got the errors (because the code detecting english is supported
didn't always run). So the easiest way to fix is just not to warn.