Add filtering controls!
This will require some sort of channel to pass query parameters outside of
the webfeed URL. And it'll require a |uniqsort filter and "in" operator.

May be interesting to add a JavaScript button that'll append the next RSS pages
to the entries list until it finds one that has some content.

But otherwise I think this draft is pretty much complete!