Fix regressions in end-user features.
I was warned that my code for intercepting form submissions will break,
but before when I tried to fix it the given solution just did work yet.
Now it does.

That impacted odysseus:home interactions.

Also viewsource broke. In part this seemed to be due to a change in how
URLs are parsed (that was standardized only recently), but it was also
due to a change in my Prosody templating language in regards to a
long-standing quirk.

Quotes in comments are lexed as strings, and I disabled support for
multiline strings for the sake of debugging. For this reason I avoid
's and "s in comments.