Occasionally beg for money.
On the one-hand I absolutely didn't want to ruin the experience I've so carefully crafted.
So I only did this because I could easily do so unobtrusively enough.

The main trick was to encourage donations on elementary OS to come from the AppCenter,
for the sake of surfers to learn how to donate to other elementary OS apps,
whilst providing a link for distros that don't include donations/kohas in their package manager.

I ended up using a macro to tell the difference.